src/tests/api/api_v1/test_context.py
from fastapi.testclient import TestClient
from src.main import app

client = TestClient(app)


def test_search_files_and_directories():
    # Test case for the search_files_and_directories endpoint
    response = client.get("/api/v1/context/search", params={"q": "test_query"})
    assert response.status_code == 200
    # Add more assertions based on expected response


def test_get_file_structure():
    # Test case for the get_file_structure endpoint
    response = client.post("/api/v1/context/file_structure/test_dir_path")
    assert response.status_code == 200
    # Add more assertions based on expected response


def test_git_diff():
    # Test case for the git_diff endpoint
    response = client.get("/api/v1/context/git_diff")
    assert response.status_code == 200
    # Add more assertions based on expected response

# Add more test cases if needed
